파일 업로드/다운로드 구현

+++ src/main/java/com/ys/carInfo/common/vo/FileVo.java
@@ -19,7 +19,7 @@ public class FileVo implements Serializable {
 	private String	fileNm;
 	private String	fileExtNm;
 	private String	mimeTy;
-	private Integer fileSize;
+	private long 	fileSize;
 	private Integer fileDownCnt;
 	private String	caption;
 	private Integer regNo;
